[core-crypto] use rollup config from dev-tool for node and browser builds
