+++ sdk/core/core-crypto/package.json
@@ -79,6 +79,7 @@
     "tslib": "^2.0.0"
   },
   "devDependencies": {
+    "@azure/dev-tool": "^1.0.0",
     "@microsoft/api-extractor": "7.7.11",
     "@rollup/plugin-commonjs": "11.0.2",
     "@rollup/plugin-json": "^4.0.0",
